refactor: arbitrage tools to their own MCP server; degeneralize prompt
Two corrections to earlier calls of mine.

1. THE TOOLS BELONG IN THEIR OWN SERVER. I put the four arbitrage tools on
stock_price_server to avoid standing up a sixth Cloud Run service. That traded
the codebase's one-domain-per-server shape for a deployment convenience, and
deployment convenience is the cheaper thing. Arbitrage has its own curated
universe, scoring model and failure modes; it gets its own wrapper like prices,
options, fundamentals, news, microstructure and portfolio do.

  fastMCPTest/arbitrage_server.py   new, port 6006, tools moved verbatim
                                    (docstrings carry the net-vs-gross warning
                                    and the bounds) plus mcp_health_check
  stock_price_server.py             back to 29 tools, zero arbitrage refs

No new image: the wrappers all run the shared quantcore-mcp image parameterized
by SERVER_MODULE/PORT, so this is a compose service, a .mcp.json pair, a smoke
table row, and two deploy steps.

The deploy steps are deliberately NOT added to the existing five-wrapper `for`
loops — those have no existence guard, and quantcore-arbitrage won't exist
until its one-time manual creation. They follow the quantcore-portfolio
precedent instead: describe-then-deploy, else skip with a message. .mcp.json
uses the project-number URL form because the hash-style hostname isn't
knowable before the service exists.

2. THE SIDEKICK PROMPT SHOULDN'T KNOW ABOUT ARBITRAGE. The global system
prompt had a paragraph about nav.premium_discount_pct. Domain guidance belongs
in tool descriptions, where the model reads it only when the tool is in play.
Replaced with two domain-neutral principles that serve every current and future
tool: prefer a corrected figure over a headline one and say what the headline
omits; report a structured verdict or "not applicable" as the finding it is
rather than as missing data.

The net-vs-gross guard is unweakened — it still lives in the tool description
(pinned by test_arbitrage_tool_steers_the_model_to_the_net_discount), in
scan_summary omitting the gross field, and in the card's headline slot.

Tests: the wrapper suite now drives arbitrage_server, plus a test asserting the
four tools are absent from stock_price_server so they can't drift back, and one
pinning the server's own identity.

1140 backend tests pass, coverage 86.4%. Wrapper smoke 7/7 (stock-price 29,
arbitrage 5). Guards pass; OpenAPI unchanged at 105 routes — the REST tier was
not touched.
