+++ .github/workflows/deploy.yml
@@ -241,6 +241,17 @@ jobs:
           else
             echo "quantcore-portfolio service not found — first deploy is a manual one-time step; skipping image roll-out."
           fi
+      - name: Deploy quantcore-arbitrage (image-only; env preserved)
+        # Same one-time-creation caveat as quantcore-portfolio above: the first
+        # deploy sets SERVER_MODULE=fastMCPTest.arbitrage_server + PORT=6006 and
+        # must be done manually; CI only rolls images onto an existing service.
+        run: |
+          if gcloud run services describe quantcore-arbitrage --project "$PROJECT_ID" --region "$REGION" >/dev/null 2>&1; then
+            gcloud run deploy quantcore-arbitrage --project "$PROJECT_ID" --region "$REGION" \
+              --image "$REGION-docker.pkg.dev/$PROJECT_ID/$AR_REPO/quantcore-mcp:${GITHUB_SHA::7}"
+          else
+            echo "quantcore-arbitrage service not found — first deploy is a manual one-time step; skipping image roll-out."
+          fi
       - name: Update report Job image
         run: |
           gcloud run jobs update quantcore-report --project "$PROJECT_ID" --region "$REGION" \
